[fix] HoldingDto 및 PortfolioResponse에서 quantity 필드를 int에서 long으로 변경

+++ src/main/java/farmsystem/backend/domain/portfolio/dto/PortfolioResponse.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 public record PortfolioResponse(
         String stockCode,
         String stockName,
-        int quantity,
+        long quantity,
         double ratio,
         double averagePrice
 ) {
